// src/components/enums/component.enums.ts
export enum ComponentType {
  CAMERA = 'CAMERA',
  LED_DISPLAY = 'LED_DISPLAY',
  PHONE_CHARGER = 'PHONE_CHARGER',
  LAMP = 'LAMP',
  BULB = 'BULB',
  ARM_BRACKET = 'ARM_BRACKET',
  FOUNDATION = 'FOUNDATION',
  CABLE = 'CABLE',
  SENSOR = 'SENSOR',
  CONTROLLER = 'CONTROLLER',
  BATTERY = 'BATTERY',
  SOLAR_PANEL = 'SOLAR_PANEL',
  WIRING = 'WIRING',
  MOUNTING_HARDWARE = 'MOUNTING_HARDWARE',
  OTHER = 'OTHER',
}

export enum ComponentStatus {
  INSTALLED = 'INSTALLED',
  REMOVED = 'REMOVED',
  UNDER_MAINTENANCE = 'UNDER_MAINTENANCE',
  DAMAGED = 'DAMAGED',
  REPLACED = 'REPLACED',
}